requirements.txt
snowflake-snowpark-python
snowflake-connector-python[pandas]
azure-storage-blob
python-dotenv
pandas
notebook